refactor(renderer): group jobs browser and workspace ownership
