+++ apps/desktop/src/renderer/App.tsx
@@ -1,20 +1,20 @@
 import { AgentPanel } from './agents/chat/AgentPanel'
 import { CenterWorkspace, type JobListingRequest } from './components/CenterWorkspace'
-import { JobNavigator } from './components/JobNavigator'
+import { JobNavigator } from './jobs/navigator/JobNavigator'
 import { SettingsPanel } from './components/SettingsPanel'
 import { StatusBar } from './components/StatusBar'
-import { WorkbenchLayout } from './components/WorkbenchLayout'
+import { WorkbenchLayout } from './workspace/WorkbenchLayout'
 import { CareerProfileWorkspace } from './career-profile/CareerProfileWorkspace'
 import { hasCachedCareerProfile } from './career-profile/work-arrangement/useCareerProfile'
-import { WorkspaceBar, type WorkspaceBarWorkspace } from './components/WorkspaceBar'
-import { BrowseWorkspace } from './components/BrowseWorkspace'
+import { WorkspaceBar, type WorkspaceBarWorkspace } from './workspace/WorkspaceBar'
+import { BrowseWorkspace } from './jobs/browse/BrowseWorkspace'
 import { useAgentAvatarPreference } from './agents/avatar/useAgentAvatarPreference'
 import { DocxDocumentEditorShell } from './document-editor/DocxDocumentEditorShell'
 import { useConnectivity } from './hooks/useConnectivity'
-import { useJobs } from './hooks/useJobs'
+import { useJobs } from './jobs/useJobs'
 import { useAgentSessions } from './agents/chat/useAgentSessions'
 import { useConnectedAgents } from './agents/connected-agents/useConnectedAgents'
-import { useWorkspace } from './hooks/useWorkspace'
+import { useWorkspace } from './workspace/useWorkspace'
 import { useTheme } from './theme/useTheme'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import type { DocxOpenResult } from '../shared/docxDocuments'
