+++ apps/desktop/src/renderer/workspace/WorkspaceBar.test.tsx
