+++ apps/desktop/src/renderer/workspace/WorkspaceBar.tsx
@@ -1,5 +1,5 @@
 import { Moon, RotateCcw, Sun } from 'lucide-react'
-import type { TopLevelWorkspace } from '../workspaceLayout'
+import type { TopLevelWorkspace } from './workspaceLayout'
 import type { ThemeMode } from '../theme/themes'
 import { InstallationProfileMenu } from '../installation-profiles/InstallationProfileMenu'
