+++ apps/desktop/src/renderer/workspace/workspaceSizing.test.ts
